+++ InfoPanel.SteamAPI/Services/Sensors/PlayerSensorService.cs
@@ -233,14 +233,13 @@ namespace InfoPanel.SteamAPI.Services.Sensors
             }
             else
             {
-                _currentGameSensor.Value = "Not Playing";
-
-                // If we have a last played game, show its playtime
+                // If we have a last played game, show its name and playtime
                 if (!string.IsNullOrEmpty(playerData.LastPlayedGameName))
                 {
+                    _currentGameSensor.Value = playerData.LastPlayedGameName;
                     _currentGamePlaytimeSensor.Value = (float)Math.Round(playerData.LastPlayedGamePlaytimeHours, 1);
 
-                    _enhancedLogger?.LogDebug($"{DOMAIN_NAME}.UpdateCurrentGameSensors", "Not playing - showing last played game playtime", new
+                    _enhancedLogger?.LogDebug($"{DOMAIN_NAME}.UpdateCurrentGameSensors", "Not playing - showing last played game info", new
                     {
                         LastPlayedGame = playerData.LastPlayedGameName,
                         TotalPlaytimeHours = Math.Round(playerData.LastPlayedGamePlaytimeHours, 1)
@@ -248,6 +247,7 @@ namespace InfoPanel.SteamAPI.Services.Sensors
                 }
                 else
                 {
+                    _currentGameSensor.Value = "Not Playing";
                     _currentGamePlaytimeSensor.Value = 0f;
 
                     _enhancedLogger?.LogDebug($"{DOMAIN_NAME}.UpdateCurrentGameSensors", "Not playing - sensors set to default");